+++ sites/playground/server/src/a2a-tools/invoke-agent/index.ts
@@ -1,1 +1,3 @@
 export * from './invoke-a2a-agent.js';
+export * from './request/index.js';
+export * from './send-message/index.js';
