+++ sites/playground/server/src/a2a-tools/invoke-agent/send-message/adapters/index.ts
@@ -1,0 +1,2 @@
+export * from './v0_3.js';
+export * from './v1_0.js';
